Initial committ of testing code
The point of this code is to make it easier to query the NCS repository
for faceted searches. This code reads the XSD document in the NCS
repository, as well as its field lists, and uses this information to
formulae queries. Very little (hopefully nothing) is being hard coded,
so if the NCS's XSD changes (as metadata requirements in the repository
evolve), this will still work and allow for queries against the new
metadata schema.