JSON representation fixes, new linking rel for Atom
I fixed the JSON representation (and PHP array object) to allow for
multiple values for metadata elements in query result items. It makes a
somewhat more complicated JSON representation, but one that better fits
the metadata schema. I also added a "describedby"linking relation to
entries in the Atom feed of query results. This links to the NCS XML
metadata descrption for a result entry.